cwl_bot/requirements.txt
discord.py
supabase
python-dotenv
requests
flask
